+++ src/logic/completionProvider.ts
@@ -1,6 +1,6 @@
 import { CompletionItem, CompletionItemKind, SnippetString, MarkdownString } from "vscode";
-import { LANGS } from "./constants";
-import { completions as pageLinkCompletions, iconCompletions } from "./workspaceFileLoader";
+import { LANGS, TAG_NAMES } from "./constants";
+import { pageIdCompletions, iconCompletions } from "./workspaceFileLoader";
 
 export const getInpageLinkCompletions = (document: string) => {
     const inpagecompletions: CompletionItem[] = []
@@ -16,28 +16,28 @@ export const getInpageLinkCompletions = (document: string) => {
 }
 
 export const getPageLinkCompletions = () => {
-    return pageLinkCompletions;
+    return pageIdCompletions;
 }
 
 export const getIconCompletions = () => {
-	return iconCompletions
+	return iconCompletions;
 }
 
 export const getSnippetCompletions = (): CompletionItem[] => {
     const pagelinkCompletion = new CompletionItem('page_link');
-    pagelinkCompletion.insertText = new SnippetString('{% page_link $1${2: linkText=\"${3:(optional)}\"} %} $0');
+    pagelinkCompletion.insertText = new SnippetString('{% page_link $1${2: linkText=\"${3:(optional)}\" %} $0');
     pagelinkCompletion.documentation = new MarkdownString("Inserts a snippet that lets you link.");
     pagelinkCompletion.preselect = true;
     pagelinkCompletion.sortText = 'AtcmplSnippet';
     
     const inPagelinkCompletion = new CompletionItem('inpage_link');
-    inPagelinkCompletion.insertText = new SnippetString('{% inpage_link $1${2: linkText=\"${3:(optional)}\"} %} $0');
+    inPagelinkCompletion.insertText = new SnippetString('{%'+`${TAG_NAMES.INPAGE_LINK}`+'$1${2: linkText=\"${3:(optional)}\"} %} $0');
     inPagelinkCompletion.documentation = new MarkdownString("Inserts a snippet that lets you link to an anchor on the current page.");
     inPagelinkCompletion.preselect = true;
     inPagelinkCompletion.sortText = 'AtcmplSnippet';
 
     const codeCompletion = new CompletionItem('code_tag');
-    codeCompletion.insertText = new SnippetString('{% code lang=${1|'+`${LANGS.join(',')}`+'|} title=\"$2\"${3: linenumbers=true} %}\r\n$0\r\n{% endcode %}\r\n\r\n');
+    codeCompletion.insertText = new SnippetString('{% '+`${TAG_NAMES.CODE}`+' lang=${1|'+`${LANGS.join(',')}`+'|} title=\"$2\"${3: linenumbers=true} %}\r\n$0\r\n{% endcode %}\r\n\r\n');
     codeCompletion.documentation = new MarkdownString("Inserts a snippet that lets you link to an anchor on the current page.");
     codeCompletion.preselect = true;
     codeCompletion.sortText = 'AtcmplSnippet';
@@ -83,6 +83,24 @@ export const getSnippetCompletions = (): CompletionItem[] => {
     tipCompletion.documentation = new MarkdownString("Inserts tip admonition snippet.");
     tipCompletion.preselect = true;
     tipCompletion.sortText = 'AtcmplSnippet';
+
+    const tableCompletion = new CompletionItem('table_tag');
+    tableCompletion.insertText = new SnippetString('{% table %}\r\n$0\r\n{% entable %}');
+    tableCompletion.documentation = new MarkdownString("Inserts table start/end.");
+    tableCompletion.preselect = true;
+    tableCompletion.sortText = 'AtcmplSnippet';
+
+    const rowCompletion = new CompletionItem('row_tag');
+    rowCompletion.insertText = new SnippetString('{% row ${1: header=\"${2:(true/false)}\" %}\r\n$0\r\n{% endrow %}');
+    rowCompletion.documentation = new MarkdownString("Inserts row start/end.");
+    rowCompletion.preselect = true;
+    rowCompletion.sortText = 'AtcmplSnippet';
+
+    const cellCompletion = new CompletionItem('cell_tag');
+    cellCompletion.insertText = new SnippetString('{% cell %}$0\r\n{% endcell %}');
+    cellCompletion.documentation = new MarkdownString("Inserts cell start/end.");
+    cellCompletion.preselect = true;
+    cellCompletion.sortText = 'AtcmplSnippet';
     
     return [
         pagelinkCompletion,
@@ -94,6 +112,9 @@ export const getSnippetCompletions = (): CompletionItem[] => {
         noteCompletion,
         infoCompletion,
         warningCompletion,
-        tipCompletion
+        tipCompletion,
+        tableCompletion,
+        rowCompletion,
+        cellCompletion
     ]
 }