+++ src/logic/constants.ts
@@ -2,6 +2,8 @@ export const LANGS: string[] = ['csharp','css','cshtml','html','js', 'json','tsx
 
 export const MARKDOWN = 'markdown';
 export const XP_DOCUMENTATION_FILES = `**/src/_documentation/_xp/**`
+export const TUTORIAL_DOCUMENTATION_FILES = `**/src/_documentation/_api/**`
+export const API_DOCUMENTATION_FILES = `**/src/_documentation/_tutorial/**`
 export const XP_ICON_DEFINITIONS = '**/src/_assets/less/**/xp-icon-variables.less'
 
 export enum Scope {